move fetch out
